refactor reporting job lineage route
