+++ src/app/router_registry.py
@@ -226,6 +226,7 @@ from app.routers.reporting_batch_status import status_router as reporting_batch_
 from app.routers.reporting_batch_worker import worker_router as reporting_batch_worker_router
 from app.routers.reporting_batches import batches_router as reporting_batches_router
 from app.routers.reporting_job_controls import controls_router as reporting_job_controls_router
+from app.routers.reporting_job_lineage import lineage_router as reporting_job_lineage_router
 from app.routers.reporting_job_search import search_router as reporting_job_search_router
 from app.routers.reporting_job_submissions import router as reporting_job_submissions_router
 from app.routers.reporting_jobs import jobs_router as reporting_jobs_router
@@ -400,6 +401,7 @@ def register_routers(app: FastAPI) -> None:
     app.include_router(reporting_snapshots_router)
     app.include_router(reporting_job_search_router)
     app.include_router(reporting_jobs_router)
+    app.include_router(reporting_job_lineage_router)
     app.include_router(reporting_job_controls_router)
     app.include_router(reporting_batches_router)
     app.include_router(reporting_batch_status_router)
